requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==6.0.1
django-admin-interface==0.32.0
django-colorfield==0.14.0
djangorestframework==3.16.1
gunicorn==25.0.1
packaging==26.0
pillow==12.1.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
python-slugify==8.0.4
sqlparse==0.5.5
text-unidecode==1.3
tzdata==2025.3
whitenoise==6.11.0
